Function useful to check Gurobi's presence is introduced
A function to check solvers is introduced in CommandGovernor's constructor.  If Gurobi is not found using getavailablesolvers() then BMIBNB is used. Otherwise Gurobi is setted as the numerical solver. Notice that the function check_solver does not verify the license of the solver.

Tested several times with Distributed and Centralized  versions of CG with Gurobi solver and works. If Gurobi is not present it should work changing the solver into bmibnb