New Planner, Check function adjustment, Virtual Reference Generator
A new planner for a generic closed trajectory is added, the check function now does not need the .mat file, the Virtual Reference Generator function is implemented. A new folder for oil stain simulation is added too, and a script to test Dynamic PnP of a vehicle is being processed.